fix adding products

+++ lib/objects/product_related/cpu_related/cpu.dart
@@ -46,7 +46,7 @@ class CPU extends Product {
     String? productName,
     int? importPrice,
     int? sellingPrice,
-    double? discount,
+    int? discount,
     DateTime? release,
     int? sales,
     int? stock,
