+++ lib/objects/product_related/drive_related/drive.dart
@@ -47,7 +47,7 @@ class Drive extends Product {
     String? productName,
     int? importPrice,
     int? sellingPrice,
-    double? discount,
+    int? discount,
     DateTime? release,
     int? sales,
     int? stock,
