+++ lib/objects/product_related/psu_related/psu.dart
@@ -41,7 +41,7 @@ class PSU extends Product {
     String? productName,
     int? importPrice,
     int? sellingPrice,
-    double? discount,
+    int? discount,
     DateTime? release,
     int? sales,
     int? stock,
