+++ lib/screens/product/product_detail/product_detail_view.dart
@@ -698,7 +698,7 @@ class _ProductDetailScreenState extends State<ProductDetailScreen> {
   Widget _buildPriceSection({
     required int importPrice,
     required int sellingPrice,
-    required double discount,
+    required int discount,
   }) {
     final colorScheme = Theme.of(context).colorScheme;
     final discountedPrice = sellingPrice * (100 - discount) / 100;
